iam array mode refactor

+++ src/main/java/bee/creative/mmi/MMIArrayL2D.java
@@ -12,7 +12,7 @@ final class MMIArrayL2D extends MMIArrayL {
 	}
 
 	@Override
-	public byte mode() {
+	public int mode() {
 		return IAMArray.MODE_INT32;
 	}
